+++ PSModule/IntuneDocumentation/Functions/Invoke-IntuneDocumentation.ps1
@@ -1,6 +1,3 @@
-#Requires -Modules Microsoft.Graph.Intune,PSWord
-#Requires -Version 5
-
 <#
 .DESCRIPTION
 This Script documents an Intune Tenant with almost all settings, which are available over the Graph API.
@@ -8,7 +5,7 @@ NOTE: This no longer does Conditional Access
 The Script is using the PSWord and Microsoft.Graph.Intune Module. Therefore you have to install them first.
 
 .EXAMPLE
-.\Invoke-IntuneDocumentation.ps1
+Invoke-IntuneDocumentation.ps1 -FullDocumentationPath c:\temp\IntuneDoc.docx
 
 .NOTES
 Author: Thomas Kurth/baseVISION
@@ -25,364 +22,23 @@ ExitCodes:
     99003: Could not Set ExitMessageRegistry
 #>
 [CmdletBinding()]
-Param()
+Param(
+    [ValidateScript({
+        if($_ -notmatch "(\.docx)"){
+            throw "The file specified in the path argument must be of type docx"
+        }
+        return $true 
+    })]
+    [System.IO.FileInfo]$FullDocumentationPath = ".\IntuneDocumentation.docx"
+)
 ## Manual Variable Definition
 ########################################################
 $DebugPreference = "Continue"
-$ScriptVersion = "014"
 $ScriptName = "DocumentIntune"
-$LogFilePathFolder     = Join-Path -Path $Env:TEMP -ChildPath $ScriptName
-# Log Configuration
-$DefaultLogOutputMode  = "Console" # "Console-LogFile","Console-WindowsEvent","LogFile-WindowsEvent","Console","LogFile","WindowsEvent","All"
-$DefaultLogWindowsEventSource = $ScriptName
-$DefaultLogWindowsEventLog = "CustomPS"
 $MaxStringLengthSettings = 350
-$DocumentName = "DocumentIntune.docx"
-$DateTimeRegex = "\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\.\d{7}Z" 
-#region Functions
-########################################################
-function Write-Log {
-    <#
-    .DESCRIPTION
-    Write text to a logfile with the current time.
 
-    .PARAMETER Message
-    Specifies the message to log.
-
-    .PARAMETER Type
-    Type of Message ("Info","Debug","Warn","Error").
-
-    .PARAMETER OutputMode
-    Specifies where the log should be written. Possible values are "Console","LogFile" and "Both".
-
-    .PARAMETER Exception
-    You can write an exception object to the log file if there was an exception.
-
-    .EXAMPLE
-    Write-Log -Message "Start process XY"
-
-    .NOTES
-    This function should be used to log information to console or log file.
-    #>
-    param(
-        [Parameter(Mandatory=$true,Position=1)]
-        [String]
-        $Message
-    ,
-        [Parameter(Mandatory=$false)]
-        [ValidateSet("Info","Debug","Warn","Error")]
-        [String]
-        $Type = "Debug"
-    ,
-        [Parameter(Mandatory=$false)]
-        [ValidateSet("Console-LogFile","Console-WindowsEvent","LogFile-WindowsEvent","Console","LogFile","WindowsEvent","All")]
-        [String]
-        $OutputMode = $DefaultLogOutputMode
-    ,
-        [Parameter(Mandatory=$false)]
-        [Exception]
-        $Exception
-    )
-    $DateTimeString = Get-Date -Format "yyyy-MM-dd HH:mm:sszz"
-    $Output = ($DateTimeString + "`t" + $Type.ToUpper() + "`t" + $Message)
-    if($Exception){
-        $ExceptionString =  ("[" + $Exception.GetType().FullName + "] " + $Exception.Message)
-        $Output = "$Output - $ExceptionString"
-    }
-    if ($OutputMode -eq "Console" -OR $OutputMode -eq "Console-LogFile" -OR $OutputMode -eq "Console-WindowsEvent" -OR $OutputMode -eq "All") {
-        if($Type -eq "Error"){
-            Write-Error $output
-        } elseif($Type -eq "Warn"){
-            Write-Warning $output
-        } elseif($Type -eq "Debug"){
-            Write-Debug $output
-        } else{
-            Write-Verbose $output -Verbose
-        }
-    }
-    if ($OutputMode -eq "LogFile" -OR $OutputMode -eq "Console-LogFile" -OR $OutputMode -eq "LogFile-WindowsEvent" -OR $OutputMode -eq "All") {
-        try {
-            Add-Content $LogFilePath -Value $Output -ErrorAction Stop
-        } catch {
-            exit 99001
-        }
-    }
-    if ($OutputMode -eq "Console-WindowsEvent" -OR $OutputMode -eq "WindowsEvent" -OR $OutputMode -eq "LogFile-WindowsEvent" -OR $OutputMode -eq "All") {
-        try {
-            New-EventLog -LogName $DefaultLogWindowsEventLog -Source $DefaultLogWindowsEventSource -ErrorAction SilentlyContinue
-            switch ($Type) {
-                "Warn" {
-                    $EventType = "Warning"
-                    break
-                }
-                "Error" {
-                    $EventType = "Error"
-                    break
-                }
-                default {
-                    $EventType = "Information"
-                }
-            }
-            Write-EventLog -LogName $DefaultLogWindowsEventLog -Source $DefaultLogWindowsEventSource -EntryType $EventType -EventId 1 -Message $Output -ErrorAction Stop
-        } catch {
-            exit 99002
-        }
-    }
-}
-function New-Folder{
-    <#
-    .DESCRIPTION
-    Creates a Folder if it's not existing.
-
-    .PARAMETER Path
-    Specifies the path of the new folder.
-
-    .EXAMPLE
-    CreateFolder "c:\temp"
-
-    .NOTES
-    This function creates a folder if doesn't exist.
-    #>
-    param(
-        [Parameter(Mandatory=$True,Position=1)]
-        [string]$Path
-    )
-	# Check if the folder Exists
-	if (Test-Path $Path) {
-		Write-Log "Folder: $Path Already Exists"
-	} else {
-		New-Item -Path $Path -type directory | Out-Null
-		Write-Log "Creating $Path"
-	}
-}
-Function Get-WindowsAutopilotConfig(){
-    <#
-    .SYNOPSIS
-    This function is used to get the AutoPilot configuration from the Graph API REST interface
-    .DESCRIPTION
-    The function connects to the Graph API Interface and gets the AutoPilot Configuration
-    .EXAMPLE
-    Get-WindowsAutopilotConfig
-    Returns the AutoPilot Configuration configured in Intune
-    .NOTES
-    NAME: Get-WindowsAutopilotConfig
-    #>
-    try {
-        $uri = "https://graph.microsoft.com/Beta/deviceManagement/windowsAutopilotDeploymentProfiles"
-        (Invoke-MSGraphRequest -Url $uri -HttpMethod GET).Value
-    } catch {
-        $ex = $_.Exception
-        $errorResponse = $ex.Response.GetResponseStream()
-        $reader = New-Object System.IO.StreamReader($errorResponse)
-        $reader.BaseStream.Position = 0
-        $reader.DiscardBufferedData()
-        $responseBody = $reader.ReadToEnd();
-        Write-Log "Response content:`n$responseBody" -Type Error
-        Write-Log "Request to $Uri failed with HTTP Status $($ex.Response.StatusCode) $($ex.Response.StatusDescription)" -Type Error
-        
-    }
-}
-Function Get-EnrollmentStatusPage(){
-<#
-    .SYNOPSIS
-    This function is used to get the Enrollment Status Page configuration from the Graph API REST interface
-    .DESCRIPTION
-    The function connects to the Graph API Interface and gets the Enrollment Status Page Configuration
-    .EXAMPLE
-    Get-EnrollmentStatusPage
-    Returns the Enrollment Status Page Configuration configured in Intune
-    .NOTES
-    NAME: Get-EnrollmentStatusPage
-    #>
-
-    try {
-        $uri = "https://graph.microsoft.com/Beta/deviceManagement/deviceEnrollmentConfigurations"
-        (Invoke-MSGraphRequest -Url $uri -HttpMethod GET).Value
-    } catch {
-        $ex = $_.Exception
-        $errorResponse = $ex.Response.GetResponseStream()
-        $reader = New-Object System.IO.StreamReader($errorResponse)
-        $reader.BaseStream.Position = 0
-        $reader.DiscardBufferedData()
-        $responseBody = $reader.ReadToEnd();
-        Write-Log "Response content:`n$responseBody" -Type Error
-        Write-Log "Request to $Uri failed with HTTP Status $($ex.Response.StatusCode) $($ex.Response.StatusDescription)" -Type Error
-    }
-}
-Function Get-DeviceManagementScripts(){
-    <#
-        .SYNOPSIS
-        This function is used to get the Intune PowerShell Scripts from the Graph API REST interface
-        .DESCRIPTION
-        The function connects to the Graph API Interface and gets the Intune PowerShell Scripts including the scripts
-        .EXAMPLE
-        Get-DeviceManagementScripts
-        Returns the Enrollment Status Page Configuration configured in Intune
-        .NOTES
-        NAME: Get-DeviceManagementScripts
-        #>
-    
-        try {
-            $uri = "https://graph.microsoft.com/Beta/deviceManagement/deviceManagementScripts"
-            $request= (Invoke-MSGraphRequest -Url $uri -HttpMethod GET)
-            $allScripts= @()
-
-            $request.value.GetEnumerator() | ForEach-Object {
-
-                $currentScript =Invoke-MSGraphRequest -HttpMethod GET -Url "https://graph.microsoft.com/Beta/deviceManagement/deviceManagementScripts/$($PSItem.id)"
-                
-                $allScripts += [PSCustomObject]@{
-                    id = $currentScript.id
-                    displayName = $currentScript.displayName
-                    description = $currentScript.description
-                    enforceSignatureCheck = $PSItem.enforceSignatureCheck
-                    runAs32Bit = $PSItem.runAs32Bit
-                    runAsAccount = $PSItem.runAsAccount
-                    fileName = $PSItem.fileName
-                    scriptContent = [System.Text.Encoding]::UTF8.GetString([System.Convert]::FromBase64String($currentScript.scriptContent))
-                }
-            }
-            $allScripts
-        } catch {
-            $ex = $_.Exception
-            $errorResponse = $ex.Response.GetResponseStream()
-            $reader = New-Object System.IO.StreamReader($errorResponse)
-            $reader.BaseStream.Position = 0
-            $reader.DiscardBufferedData()
-            $responseBody = $reader.ReadToEnd();
-            Write-Log "Response content:`n$responseBody" -Type Error
-            Write-Log "Request to $Uri failed with HTTP Status $($ex.Response.StatusCode) $($ex.Response.StatusDescription)" -Type Error
-        }
-}
-Function Get-ConditionalAccess(){
-    <#
-    .SYNOPSIS
-    This function is used to get the Conditional Access from the Graph API REST interface
-    .DESCRIPTION
-    The function connects to the Graph API Interface and gets the Azure AD Conditional Access Configuration
-    .EXAMPLE
-    Get-ConditionalAccess
-    Returns the Conditional Access Configuration configured in AzureAD
-    .NOTES
-    NAME: Get-ConditionalAccess
-    #>
-    try {
-        $uri = "https://graph.microsoft.com/Beta/conditionalAccess/policies"
-        (Invoke-MSGraphRequest -Url $uri -HttpMethod GET).Value
-    } catch {
-        $ex = $_.Exception
-        $errorResponse = $ex.Response.GetResponseStream()
-        $reader = New-Object System.IO.StreamReader($errorResponse)
-        $reader.BaseStream.Position = 0
-        $reader.DiscardBufferedData()
-        $responseBody = $reader.ReadToEnd();
-        Write-Log "Response content:`n$responseBody" -Type Error
-        Write-Log "Request to $Uri failed with HTTP Status $($ex.Response.StatusCode) $($ex.Response.StatusDescription)" -Type Error
-        
-    }
-}
-
-Function Get-ADMXBasedConfigurationProfiles(){
-    <#
-    .SYNOPSIS
-    This function is used to get the ADMX Policies from the Graph API REST interface
-    .DESCRIPTION
-    The function connects to the Graph API Interface and gets the Intune ADMX Configuration
-    .EXAMPLE
-    Get-ADMXBasedConfigurationProfiles
-    Returns the ADMX based Configuration Profiles configured in Intune
-    .NOTES
-    NAME: Get-ADMXBasedConfigurationProfiles
-    #>
-    try {
-        Update-MSGraphEnvironment -SchemaVersion 'beta'
-        Connect-MSGraph
-        $Policies = Invoke-MSGraphRequest -HttpMethod GET -Url "/deviceManagement/groupPolicyConfigurations"
-        $return = @()
-        foreach($Policy in $Policies.value){
-            $return2 = @()
-            $values = Invoke-MSGraphRequest -HttpMethod GET -Url "/deviceManagement/groupPolicyConfigurations/$($Policy.Id)/definitionValues"
-            foreach($value in $values.value){
-                try{
-                    $definition = (Invoke-MSGraphRequest -HttpMethod GET -Url "/deviceManagement/groupPolicyConfigurations/$($Policy.Id)/definitionValues/$($value.id)/definition")
-                    $res = Invoke-MSGraphRequest -HttpMethod GET -Url "/deviceManagement/groupPolicyConfigurations/$($Policy.Id)/definitionValues/$($value.id)/presentationValues"
-                    $return2 += [PSCustomObject]@{ 
-                        DisplayName = $definition.displayName
-                        #ExplainText = $definition.explainText
-                        Scope = $definition.classType
-                        Path = $definition.categoryPath
-                        SupportedOn = $definition.supportedOn
-                        Enabled = $value.enabled
-                        Value = if($res.value.value.GetType().baseType.Name -eq "Array"){ $res.value.value -join ", "  }else { $res.value.value }
-                    }
-                } catch {
-
-                }
-            }
-            $return += [PSCustomObject]@{ 
-                DisplayName = $Policy.displayName
-                Settings = $return2
-            }
-        }
-        Update-MSGraphEnvironment -SchemaVersion v1.0
-        Connect-MSGraph
-        return $return
-    } catch {     
-        Write-Log "Response content:`n$responseBody" -Type Error
-        Write-Log "Failed to get ADMX based Intune Policies." -Type Error -Exception $_.Exception
-    }
-}
-
-Function Format-MsGraphData(){
-    <#
-    .SYNOPSIS
-    This function Cleansup Values Returned By Microsoft Graph
-    .DESCRIPTION
-    This function Cleansup Values Returned By Microsoft Graph
-    .EXAMPLE
-    Format-MsGraphData -Value "@Odata.Type"
-    Returns "Type"
-    .NOTES
-    NAME: Format-MsGraphData
-    #>
-    [cmdletbinding()]
-    param
-    (
-        [Parameter(Mandatory=$false)]
-        [AllowEmptyString()]
-        [AllowNull()]
-        $Value
-    )
-    if($value -is [array]){
-        $value = $value -join ","
-    }
-    [string]$value = "$value"
-    $Value = $Value -replace "#microsoft.graph.",""
-    $Value = $Value -replace "windows","win"
-    $Value = $Value -replace "StoreforBusiness","SfB"
-    $Value = $Value -replace "@odata.",""
-    if($null -ne $Value -and $Value -match "@{*"){
-        $Value = $Value -replace "@{",""
-        $Value = $Value -replace "}",""
-        $Value = $Value -replace ";",""
-    }
-    if($Value -match $DateTimeRegex){
-        try{
-            [DateTime]$Date = ([DateTime]::Parse($Value))
-            $Value = "$($Date.ToShortDateString()) $($Date.ToShortTimeString())"
-        } catch {
-        }
-    }
-    return $value
-}
-#endregion
-#region Dynamic Variables and Parameters
-########################################################
-$LogFilePath = "$LogFilePathFolder\{0}_{1}_{2}.log" -f ($ScriptName -replace ".ps1", ''),$ScriptVersion,(Get-Date -uformat %Y%m%d%H%M)
-#endregion
 #region Initialization
 ########################################################
-New-Folder $LogFilePathFolder
 Write-Log "Start Script $Scriptname"
 #region Authentication
 Connect-MSGraph
@@ -391,33 +47,13 @@ Connect-MSGraph
 #region Main Script
 ########################################################
 #region Save Path
-try{
-    $SaveFileDialog = New-Object windows.forms.savefiledialog
-    $SaveFileDialog.initialDirectory = $LogFilePathFolder 
-    $SaveFileDialog.title = "Save File to Disk (If File exists, content will be appended)"   
-    $SaveFileDialog.filter = "Word Document (*.docx)|*.docx" 
-    $SaveFileDialog.ShowHelp = $True   
-    Write-Log "Where would you like to create documentation file?... (see File Save Dialog)"
-    $result = $SaveFileDialog.ShowDialog()    
-    if($result -eq "OK")    {    
-        Write-Log "Selected File and Location: $($SaveFileDialog.filename )" 
-        $FullDocumentationPath = $SaveFileDialog.filename
-    } 
-    else { 
-        Write-Log "File Save Dialog Cancelled! Using Default Path: $LogFilePathFolder\$DocumentName" -Type Warn
-        $FullDocumentationPath = "$LogFilePathFolder\$DocumentName"
-    } 
-    $SaveFileDialog.Dispose()
-} catch {
-    Write-Log "File Save Dialog Cancelled! Using Default Path: $LogFilePathFolder\$DocumentName" -Type Warn
 
-}
 #endregion
 #region CopyTemplate
 if((Test-Path -Path $FullDocumentationPath)){
     Write-Log "File already exists, does not use built-in template." -Type Warn
 } else {
-    Copy-Item "$PSScriptRoot\Template.docx" -Destination $FullDocumentationPath
+    Copy-Item "$PSScriptRoot\..\Data\Template.docx" -Destination $FullDocumentationPath
     Update-WordText -FilePath $FullDocumentationPath -ReplacingText "DATE" -NewText (Get-Date -Format "HH:mm dd.MM.yyyy")
     try{
         $org = Invoke-MSGraphRequest -Url /organization
@@ -591,7 +227,7 @@ foreach($DCP in $DCPs){
         }
     }
 }
-$ADMXPolicies = Get-ADMXBasedConfigurationProfiles
+$ADMXPolicies = Get-ADMXBasedConfigurationProfile
 foreach($ADMXPolicy in $ADMXPolicies){
     write-Log "Device Configuration (ADMX): $($ADMXPolicy.DisplayName)"
     Add-WordText -FilePath $FullDocumentationPath -Heading Heading2 -Text $ADMXPolicy.DisplayName
@@ -600,7 +236,7 @@ foreach($ADMXPolicy in $ADMXPolicies){
 #endregion
 #region Device Management Scripts (PowerShell)
 Add-WordText -FilePath $FullDocumentationPath -Heading Heading1 -Text "Device Management Scripts"
-$PSScripts = Get-DeviceManagementScripts
+$PSScripts = Get-DeviceManagementScript
 foreach($PSScript in $PSScripts){
     Add-WordText -FilePath $FullDocumentationPath -Heading Heading2 -Text $PSScript.displayName
     $ht2 = @{}